+++ tests/utils/zephyr/lattice_embedding/test_chain.py
@@ -16,7 +16,7 @@
 
 from unittest import TestCase
 
-from minorminer.utils.lattice_embedding.chain import ZephyrVHChain
+from minorminer.utils.zephyr.lattice_embedding.chain import ZephyrVHChain
 from minorminer.utils.zephyr.node_edge import ZNode, ZShape
 from minorminer.utils.zephyr.plane_shift import PlaneShift
